backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pyinstaller>=6.4.0
